split remove peep functions
